# oefeningLaravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookstore API",
        "description": "Bookstore API",
        "contact": {
            "email": "lukas@b302.nl"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "register",
                "requestBody": {
                    "description": "Register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success: Register"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success: Login"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "logout",
                "requestBody": {
                    "description": "Logout",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success: Logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get all authors",
                "description": "Get all authors",
                "operationId": "indexAuthors",
                "requestBody": {
                    "description": "Authors",
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success: All authors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Store an author",
                "description": "Store an author, this can only be done by an authenticated user.",
                "operationId": "storeAuthors",
                "requestBody": {
                    "description": "Author",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success: An author has been created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors/{author}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get an author",
                "description": "Get an author",
                "operationId": "showAuthors",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Id of author that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: An author has been found"
                    },
                    "404": {
                        "description": "An author has not been found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an author",
                "description": "Update an author",
                "operationId": "updateAuthors",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Id of author that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Author",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success: An author has been updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete an author",
                "description": "Delete an author",
                "operationId": "destroyAuthors",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "Id of author that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: An author has been deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Author": {
                "title": "Author",
                "description": "Author model",
                "properties": {
                    "name": {
                        "title": "Author name",
                        "description": "Author name",
                        "type": "string",
                        "example": "J.K. Rowling"
                    },
                    "age": {
                        "title": "Author age",
                        "description": "Author age",
                        "type": "string",
                        "example": "55"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Author"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "example": "brad@email.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "brad123"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}